jenkins: remove envfile in cleanup post action
The application uses envfiles for production environment. Jenkins decrypts
the `.env.enc` during every build.

Envfile is decrypted after application tests, but before application
build process. This is because we don't want to test application using
production configuration with production credentials.

We had a problem with CI process because only first build successes.
The second and next builds fail because `.env` is not removed in cleanup
post action.

This commit fixes this problem.