[CM2-1334] Fix homeowner account creation (#758)
The regex used for sanitizing the strings was removing the @ and . chars
from the email address, that was causing the error.

The fix was to not sanitize the email string. Email validation/sanitization
is tricky to get it right, soon we're moving to a new implementation, in
which the back-end may take of sanitization of this.